Move web3status
